Ensure users cannot share an email address
This commit provides changes to the user management creation and
update logic to enforce that users cannot share an email address.

Currently this constraint is not enforced at the database level.

A migration has been added to fix potentially inconsistent databases,
by updating the email value in the db for users that share an email
address. The migration logs information about the updated users in the
service log, so that an administrator can take action and fix potential
issues (this is not optimal, but good enough for now).

Issue: https://issues.infn.it/jira/browse/INDIAM-125