Just-in-time account provisioning
JIT account provisioning can be configured when saml external
authentication support is enabled, so that no registration is necessary
and an account is created on demand after SAML authentication based on
the information returned in the assertion.

It is also possible to configure periodic account cleanup so that
inactive accounts are cleaned up after a configurable period of
inactivity.