X.509 authentication and account linking
This commit introduces working native X.509 authentication. Users can
link certificates they own to their membership. Administrators can add
certificates to users by providing a PEM-encoded certificate. The IAM
checks that the certificate, when first added, is not already bound to a
different user.
The certificate subject and issuer, in RFC2253 format, are stored in the
database. The PEM encoded certificate is stored as well.