Improved X.509 session management
This commit introduces improved X.509 session management to the IAM.
User authenticated with an X.509 certificate linked to an IAM account
are not *automatically* logged in as before. Now the user has to
explicitly request that it wants to authenticate with an X.509
certificate, and the logout button works as expected.

The login page has been refactored to explicitly include information
about X.509 authentication (when present).