Fix MySQL migrations & other stuff
- Fixed migrations for MySQL to support mitre 1.3.x
- Improved external service health endpoint so that it can accomodate
any URL for the checks and can be disabled
- Tests that check external connectivity can now be disabled by setting
the iam.offline=true system property (useful when testing without
connectivity)
- Added a smarter date equals matcher: H2 stores milliseconds in dates
persisted to the database. MySQL stores up to the second with our
current schema. A new date equals matcher that takes this into account
is added and used in tests that failed when run against MySQL.
- Added the WithAnonymousUser annotation that should be used explicitly
in the initialization of all integration tests