Support for external authentication hint
There are scenarios where RPs want to directly hint at which authn
mechanism should be used by a user when logging in.

As an example, an application may have a "Login with INFN" button that will
redirect the user to the IAM requesting that the user is then redirected to
the INFN SAML IdP for authentication, without showing the usual IAM
login page.

This commit introduces initial support for this functionality for
external SAML IdPs. An hint-aware spring security authentication entry
point is added that is run when an authentication flow needs to start.
This entry point checks if the request was an authorize request and if a
parameter, `ext_authn_hint` was provided. If so, the requested authn
mechanims is resolved from the hint using the
ExternalAuthenticationHintService, that crafts the URL where the user
will be redirected for authentication.

The test client application has been modified to support the
customization of the authorize request with the `ext_authn_hint`
parameter.