docker: get branch name from env variable
The Jenkins git plugin fetches a specific commit when checking out code
for a branch, so the git command used to resolve the branch name does
not work as expected in a Jenkins build (the git repo is always in
detached head state).

Jenkins provides the name of the branch in the BRANCH_NAME
environment variable in multibranch build, so the build-prod-image.sh
script has now been modified to take advantage of this.