Fix migration that breaks deployment on MySQL 5.7
Due to stricter policy on timestamp default values, IAM database
deployment fails on MySQL 5.7.

To make 1.5 installable without hassles on MySQL 5.7 we change the
migration that failed. This approach has the drawback that existing
installations on MySQL < 5.7 will not have a clean upgrade to 1.5.0. This
is fixed by including the flyway-repair profile in the prod profile
for this release.

Issue: https://github.com/indigo-iam/iam/issues/265