Consistent session timeout handling
The IAM session timeout was driven by the external authentication
token timeout, but this can lead to issues especially when the
token timeout is set really short (as in the case of the CERN SSO,
where the id token timeout is 60 seconds).

So now IAM enforces the session timeout set by the configuration across
all authentication mechanisms.