Use correct scope information in userinfo endpoint
This commit fixes a problem that leads to exposing too much information
on the userinfo endpoint.

The problems is caused by IAM using the set of scopes linked to the
original authorization request instead of the actual scopes linked to
the access token presented to the userinfo endpoint. These two sets of
scopes may differ for refreshed tokens where only a subset of the
original scopes is requested.

Issue: https://github.com/indigo-iam/iam/issues/348