Enforce intermediate group membership
IAM up to now did not enforce intermediate group membership, leaving
that task to administrators.

With this commit we introduce changes that ensure that when a member is
added to a group g, such member is also added to all the groups in the
chain from g to the root.

Also, when a member is removed from group g, it is also automatically
removed from all g subgroups.

This commit implements changes on the API. The dashboard is a bit broken
by these changes and will be fixed in another commit.