Forbid same-client token exchange requesting offline_access scope
Token exchange in IAM is implemenented mainly to provide a mechanism to
allow controlled delegation of privileges across different client
applications.

There are however clients (e.g., htgettoken) that leverage the token
exchange grant to implement token attenuation.

With this commit we introduce a check to forbid requesting the
offline_access scope in a single-client token exchange, as that kind of
request could be used to extend indefinetely the lifetime of an
exchanged refresh token.