Include wlcg.groups information in userinfo response
Even though the IAM access token is a JWT and even though groups are
included in the access token when requested, as mandated by the WLCG JWT
profile, there are still apps treating the access token as an opaque
string.

To support those apps, and be more consistent with the traditional IAM
profile behaviour, IAM should include group information in the userinfo
endpoint response also for the WLCG profile.

Issue: #432