Fix total number of tokens
shown in IAM dashboard (on the left side) when logging in.
The wrong number was due to the fact that, in addition to access tokens, registration access tokens were also considered.
By checking that the expiration date of the access tokens when querying the api is NOT NULL,
the problem has been solved. The registration access tokens never expire and are therefore excluded from the count.