Introduce new admin scopes in order to access IAM API endpoints (#562)
This PR avoids that token created by IAM admin have full access to IAM API (issue #543). In fact, the authorisation at IAM API endpoints is based on user's ROLE. The identity of the user is evaluated from browser session or from token sub value. This means that scopes are ignored. This fix introduces the fact that authorisation is based on the user's identity only if it's a browser session, otherwise only scopes count. Two new scopes have been added:
- `iam:admin.read`: it gives the same read privileges of an administrator;
- `iam:admin.write`: it gives the same read & write privileges of an administrator.